+++ Modern JavaScript From The Beginning/Section 8: API Projects/78 - 81. WeatherJS/app.js
@@ -1,12 +1,19 @@
 const defaultCountry = 'Norway';
 
 const countryAPI = new Country(defaultCountry);
+const ui = new UI();
 
 function getCountryInfo () {
     countryAPI
         .getCountryInfo()
         .then(
-            data => console.log(data)
+            data => {
+                if (data.status === 404) {
+                    throw new Error(data.message);
+                } else {
+                    ui.displayCountryInfo(data[0]);
+                }
+            }
         )
         .catch(
             error => console.log(error)